+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Edit the section for your next version **before** pushing the release tag like v
 
 
 ## vCurrent 
+
 ### Changed  <!-- Automatically added comment messages comments -->
 - Update changelog for v0.2.22
 - Bump version to 0.2.22