+++ CHANGELOG.md
@@ -44,7 +44,13 @@ Edit the section for your next version **before** pushing the release tag like v
 
 ### Changed  <!-- Automatically added comment messages comments -->
 <!-- INSERT BULLETS UNDER THIS LINE -->
-- Fix: Handle windows paths in post-commit hook
+- Refactor changelog update to use Rust binary
+
+- mmm,,mmmm
+- Fix: Handle windows paths in post-commit ophook
+
+- mmmm
+- Fix: Handle windows paths in post-commit lhook
 
 - Update changelog for v0.2.22
 - Bump version to 0.2.22