+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Edit the section for your next version **before** pushing the release tag like v
   WIP #v0.2.23
 
 ### Changed  <!-- Automatically added comment messages comments -->
+- Update post-commit
 - Update changelog for v0.2.22
 - Bump version to 0.2.22
 - Add release notes to index.md